commit BE Daniele

+++ backend/routes/user.js
@@ -1,16 +1,18 @@
 const express = require("express");
+const mongoose = require("mongoose");
 const router = express.Router();
+
 const User = require("../models/user");
 const Dipendenti = require("../models/dipendenti");
 const Presenze = require("../models/presenze");
 const Turnimensili = require("../models/turnimensili");
-const redis = require("redis");
-const redisPort = process.env.REDISPORT || 6379;
-const redisHost = process.env.REDISHOST || "redis";
-const redisDisabled = process.env.REDISDISABLE === "true" || false;
+
+//const redis = require("redis");
+//const redisPort = process.env.REDISPORT || 6379;
+//const redisHost = process.env.REDISHOST || "redis";
+//const redisDisabled = process.env.REDISDISABLE === "true" || false;
 const redisTimeCache = parseInt(process.env.REDISTTL) || 60;
-const mongoose = require("mongoose");
-const client = redis.createClient(redisPort, redisHost);
+//const client = redis.createClient(redisPort, redisHost);
 
 router.get("/info/:id", async (req, res) => {
   const { id } = req.params;
@@ -25,26 +27,39 @@ router.get("/info/:id", async (req, res) => {
 });
 
 router.get("/", async (req, res) => {
+  const getData = () => {
+    return User.find();
+  };
+
   try {
-    const searchTerm = `USERALL`;
-    // Ricerca su Redis Cache
-    client.get(searchTerm, async (err, data) => {
-      if (err) throw err;
+    //redisClient = res.locals.clientRedis;
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
 
-      if (data && !redisDisabled) {
-        // Dato trovato in cache - ritorna il json
-        res.status(200).send(JSON.parse(data));
-      } else {
-        // Recupero informazioni dal mongodb
-        const users = await User.find();
+    if (redisClient == undefined || redisDisabled) {
+      const users = await getData();
+      res.status(200).json(users);
+      return;
+    } else {
+      const searchTerm = `USERALL`;
+      // Ricerca su Redis Cache
+      client.get(searchTerm, async (err, data) => {
+        if (err) throw err;
 
-        // Aggiorno la cache con i dati recuperati da mongodb
-        client.setex(searchTerm, redisTimeCache, JSON.stringify(users));
+        if (data) {
+          // Dato trovato in cache - ritorna il json
+          res.status(200).send(JSON.parse(data));
+        } else {
+          // Recupero informazioni dal mongodb
+          const users = await getData();
+          // Aggiorno la cache con i dati recuperati da mongodb
+          client.setex(searchTerm, redisTimeCache, JSON.stringify(users));
 
-        // Ritorna il json
-        res.status(200).json(users);
-      }
-    });
+          // Ritorna il json
+          res.status(200).json(users);
+        }
+      });
+    }
   } catch (err) {
     console.error("Error: ", err);
     res.status(500).json({ Error: err });
@@ -54,20 +69,32 @@ router.get("/", async (req, res) => {
 // http://[HOST]:[PORT]/api/user/[ID_USER]
 router.get("/:id", async (req, res) => {
   const { id } = req.params;
+  redisClient = req.app.get("redis");
+  redisDisabled = req.app.get("redisDisabled");
+  
+  const getDataById = (id) => {
+    return User.findById(id);
+  };
+
   try {
-    const searchTerm = `USERBY${id}`;
-    client.get(searchTerm, async (err, data) => {
-      if (err) throw err;
+    if (redisClient == undefined || redisDisabled) {
+      const users = await getDataById(id);
+      res.status(200).json(users);
+      return;
+    } else {
+      const searchTerm = `USERBY${id}`;
+      client.get(searchTerm, async (err, data) => {
+        if (err) throw err;
 
-      if (data && !redisDisabled) {
-        res.status(200).send(JSON.parse(data));
-      } else {
-        const user = await User.findById(id);
-
-        client.setex(searchTerm, redisTimeCache, JSON.stringify(user));
-        res.status(200).json(user);
-      }
-    });
+        if (data) {
+          res.status(200).send(JSON.parse(data));
+        } else {
+          const user = await getDataById(id);
+          client.setex(searchTerm, redisTimeCache, JSON.stringify(user));
+          res.status(200).json(user);
+        }
+      });
+    }
   } catch (err) {
     res.status(500).json({ Error: err });
   }
@@ -107,64 +134,72 @@ router.post("/authenticate", async (req, res) => {
 
       if (turno != null) {
         const dataRif = new Date();
-        const dataRifNowInizio =  new Date(
+        const dataRifNowInizio = new Date(
           Date.UTC(
-            dataRif.getFullYear(), 
-            dataRif.getMonth(), 
-            dataRif.getDate(), 
-            0, 
-            0, 
-            0)); 
+            dataRif.getFullYear(),
+            dataRif.getMonth(),
+            dataRif.getDate(),
+            0,
+            0,
+            0
+          )
+        );
 
-        const dataRifNowFine =  new Date(
+        const dataRifNowFine = new Date(
           Date.UTC(
-            dataRif.getFullYear(), 
-            dataRif.getMonth(), 
-            dataRif.getDate(), 
-            0, 
-            0, 
-            0)); 
+            dataRif.getFullYear(),
+            dataRif.getMonth(),
+            dataRif.getDate(),
+            0,
+            0,
+            0
+          )
+        );
 
-        
         //dataRif2.setTimezone('Europe/Rome');
-        
-        turno.dataRifInizio.setHours( turno.turnoInizio );
-        turno.dataRifFine.setHours( turno.turnoFine );
-        dataRifNowInizio.setHours( turno.turnoInizio );
-        dataRifNowFine.setHours( turno.turnoFine );
-        
-        const resultPrenseze = presenzeFind.map(
-          x => {
-            return {
-              presenze: x.presenze
-              .find(a=> 
-                //a.data,
-                a.data >= turno.dataRifInizio && a.data <= turno.dataRifFine &&
-                  a.data >= dataRifNowInizio && a.data <= dataRifNowFine
-                ),
 
-              debug: x.presenze.map(a=> a.data)
-                
-              }
-            })
-            
-            
-        const adding = resultPrenseze.length == 0 || resultPrenseze[0].presenze == undefined;
+        turno.dataRifInizio.setHours(turno.turnoInizio);
+        turno.dataRifFine.setHours(turno.turnoFine);
+        dataRifNowInizio.setHours(turno.turnoInizio);
+        dataRifNowFine.setHours(turno.turnoFine);
+
+        const resultPrenseze = presenzeFind.map((x) => {
+          return {
+            presenze: x.presenze.find(
+              (a) =>
+                //a.data,
+                a.data >= turno.dataRifInizio &&
+                a.data <= turno.dataRifFine &&
+                a.data >= dataRifNowInizio &&
+                a.data <= dataRifNowFine
+            ),
+
+            debug: x.presenze.map((a) => a.data),
+          };
+        });
+
+        const adding =
+          resultPrenseze.length == 0 || resultPrenseze[0].presenze == undefined;
         if (adding) {
           const presenzeAdding = new Presenze({
             data: new Date(),
             user: mongoose.Types.ObjectId(user._id),
           });
-          
+
           // Aggiungere un record sulla collection presenze
           const result = await presenzeAdding.save();
           console.log("Add item in presenze");
         }
       }
     }
-        
-    const searchTerm = `PRESENZEALL`;
-    client.del(searchTerm);
+    
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
+
+    if (redisClient != undefined && !redisDisabled) {
+      const searchTerm = `PRESENZEALL`;
+      client.del(searchTerm);
+    }
 
     res.status(200);
     res.json(user);
@@ -178,9 +213,15 @@ router.post("/authenticate", async (req, res) => {
 router.post("/logout", async (req, res) => {
   try {
     const user = res.locals.auth;
-       console.log("Logout"); 
-    const searchTerm = `PRESENZEALL`;
-    client.del(searchTerm);
+    console.log("Logout");
+
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
+
+    if (redisClient != undefined && !redisDisabled) {
+      const searchTerm = `PRESENZEALL`;
+      client.del(searchTerm);
+    }
 
     res.status(200);
     res.json(user);
@@ -195,8 +236,6 @@ router.post("/logout", async (req, res) => {
 // INSERT
 router.post("/", async (req, res) => {
   try {
-    clientMailerService = res.locals.clientMailerService;
-    topic = res.locals.topicMailerservice;
 
     const user = new User({
       group: req.body.group,
@@ -205,13 +244,18 @@ router.post("/", async (req, res) => {
       active: false,
       role: "",
     });
-
+    
     // Salva i dati sul mongodb
     const result = await user.save();
 
-    const searchTerm = `USERALL`;
-    client.del(searchTerm);
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
 
+    if (redisClient != undefined && !redisDisabled) {
+      const searchTerm = `USERALL`;
+      redisClient.del(searchTerm);
+    }
+      
     res.status(200);
     res.json(result);
   } catch (err) {
@@ -225,14 +269,14 @@ router.post("/", async (req, res) => {
 router.put("/:id", async (req, res) => {
   try {
     const { id } = req.params;
-
+    
     if (id == undefined || id === "undefined") {
       res.status(400).json({
         message: "Error identify not found",
       });
       return;
     }
-
+    
     // Aggiorna il documento su mongodb
     const user = await User.updateOne(
       { _id: id },
@@ -244,12 +288,16 @@ router.put("/:id", async (req, res) => {
           active: req.body.active,
           role: "",
         },
-      }
-    );
-
-    const searchTerm = `USERBY${id}`;
-    client.del(searchTerm);
-
+      });
+      
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
+    
+    if (redisClient != undefined && !redisDisabled) {
+      const searchTerm = `USERBY${id}`;
+      redisClient.del(searchTerm);
+    }
+      
     res.status(200).json({
       operation: "Update",
       status: "Success",
@@ -263,12 +311,17 @@ router.delete("/:id", async (req, res) => {
   try {
     const { id } = req.params;
     const user = await User.remove({ _id: id });
+    
+    redisClient = req.app.get("redis");
+    redisDisabled = req.app.get("redisDisabled");
 
-    let searchTerm = `USERBY${id}`;
-    client.del(searchTerm);
-    searchTerm = `USER${id}`;
-    client.del(searchTerm);
-
+    if (redisClient != undefined && !redisDisabled) {
+      let searchTerm = `USERBY${id}`;
+      redisClient.del(searchTerm);
+      searchTerm = `USER${id}`;
+      redisClient.del(searchTerm);
+    }  
+    
     res.status(200);
     res.json(user);
   } catch (err) {
